Import revision 47982 from the trunk: More FLT_MAX to DBL_MAX updates
git-svn-id: http://root.cern.ch/svn/root/branches/v5-34-00-patches@47983 27541ba8-7e3a-0410-8455-c3a389f83636

+++ tree/tree/src/TTree.cxx
@@ -5283,7 +5283,7 @@ Double_t TTree::GetMinimum(const char* columname)
       return 0;
    }
    TBranch* branch = leaf->GetBranch();
-   Double_t cmin = FLT_MAX;
+   Double_t cmin = DBL_MAX;
    for (Long64_t i = 0; i < fEntries; ++i) {
       Long64_t entryNumber = this->GetEntryNumber(i);
       if (entryNumber < 0) break;
